Home

SymArrays.jl

SymArrays.SymArr_ifsym

SymArrays.binomial_simple

SymArrays.check_contraction_compatibility

SymArrays.contract_symindex!

SymArrays.symgrp_size

SymArrays.which_symgrp

SymArrays.@symind_binomial

SymArrays.SymArr_ifsym — Method.

SymArr_ifsym(A,Nsyms) make a SymArray if there is some symmetry (i.e., any of the Nsyms are not 1)

source

SymArrays.binomial_simple — Method.

based on Base.binomial, but without negative values for n and without overflow checks (index calculations here should not overflow if the array does not have more elements than an Int64 can represent)

source

SymArrays.check_contraction_compatibility — Method.

Check if the arguments correspond to a valid contraction. Do all "static" checks at compile time.

source

SymArrays.contract_symindex! — Method.

A[iAprev,icntrct,iApost] S[iSprev,Icntrct,ISpost] res[iAprev,iApost,iSprev,Icntrct-1,ISpost]

source

SymArrays.symgrp_size — Method.

size of a single symmetric group with Nsym dimensions and size Nt per dimension

source

SymArrays.which_symgrp — Method.

return the symmetry group index and the number of symmetric indices in the group

source

SymArrays.@symind_binomial — Macro.

calculate binomial(ii+n+offset,n), equal to prod((ii+j+offset)/j, j=1:n) This shows up in size and index calculations for arrays with symmetric indices.

source