SymArrays.jl

SymArrays.SymArr_ifsym

SymArrays.binomial_simple

SymArrays.check_contraction_compatibility

SymArrays.contract_symindex!

SymArrays.ind2sub_symgrp

SymArrays.storage_type

SymArrays.symgrp_size

SymArrays.symgrp_sortedsub2ind

SymArrays.symind2ind

SymArrays.which_symgrp

SymArrays.SymArr_ifsym — Method

SymArr_ifsym(A,Nsyms) make a SymArray if there is some symmetry (i.e., any of the Nsyms are not 1)

SymArrays.binomial_simple — Method

based on Base.binomial, but without negative values for n and without overflow checks (index calculations here should not overflow if the array does not have more elements than an Int64 can represent)

SymArrays.check_contraction_compatibility — Method

Check if the arguments correspond to a valid contraction. Do all "static" checks at compile time.

SymArrays.contract_symindex! — Method

A[iAprev,icntrct,iApost] S[iSprev,Icntrct,ISpost] res[iAprev,iApost,iSprev,Icntrct-1,ISpost]

SymArrays.ind2sub_symgrp — Method

convert a linear index for a symmetric index group into a group of subindices

SymArrays.storage_type — Method

storage_type(A)

Return the type of the underlying storage array for array wrappers.

SymArrays.symgrp_size — Method

size of a single symmetric group with Nsym dimensions and size Nt per dimension

SymArrays.symgrp_sortedsub2ind — Method

calculates the linear index corresponding to the symmetric index group (i1,...,iNsym)

SymArrays.symind2ind — Method

calculates the contribution of index idim in (i1,...,idim,...,iN) to the corresponding linear index for the group

SymArrays.which_symgrp — Method

return the symmetry group index and the number of symmetric indices in the group